readme: document dependencies
In README-developer.txt, document some of the distro packages required to
build the tools and flash images. Point out issues with libusb versions.

Enhance README-user.txt to mention that the steps in README-developer.txt
typically need to be followed first, and briefly describe what's contained
in the flash images this tool handles.

Signed-off-by: Stephen Warren <swarren@nvidia.com>