Fix typo in variable name
For some reason (likely cut/paste), all the ArgParser subparsers were
stored in a variable named parser_list_configs. That variable name is at
most appropriate for one subcommand, and in fact the subcommand it would
be relevant for is not implemented by "build" but rather
"tegra-uboot-flasher". Replace this variable name with something more
appropriate and generic.

Signed-off-by: Stephen Warren <swarren@nvidia.com>