Support U-Boot's temporary output DTB path
Historically, U-Boot has generated $(obj)/u-boot.dtb. During the Kbuild
conversion, this file was moved to $(obj)/dts/dt.dtb. While this change
will be imminentely reverted in the U-Boot makefiles, it's quite easy to
support in the flasher build scripts, so we may as well. This prevents
user confusion and/or "git bisect" issues.

Signed-off-by: Stephen Warren <swarren@nvidia.com>