Use actual U-Boot BSS size for padding when flashing
In flashing mode, tegra-uboot-flasher downloads an image into RAM which
consists of the U-Boot binary, board DTB, some padding, and the image to
write into flash. The padding needs to be large enough so that the flash
image doesn't overlap with U-Boot's BSS at run-time, since the BSS is
zero'd out at boot, and stores variables that are written to at run-time.
Any overlap would cause the flash image to be corrupted.

tegra-uboot-flasher currently uses a guess of 1MB for the required
padding size. However, when enabling DFU support in U-Boot, the BSS size
balloons to well over 1MB, thus causing corruption to the flash image.
Solve this by recording the actual BSS size at build time, and using the
exact value at flashing time to calculate the padding. Add in an extra
4KB of padding just in case the DTB gets larger when adding in the
flashing commands. Previously, this was also assumed to fit into the
hard-coded 1MB of pad.

Signed-off-by: Stephen Warren <swarren@nvidia.com>