Wrap any env var settings in quotes
This allows the user to pass the following to the flashing script:

    --env bootcmd 'usb start ; dhcp zImage'

rather than having to manually escape the commands for U-Boot:

    --env bootcmd 'usb start \; dhcp zImage'

(The quoting in both cases is for the shell invoking tegra-uboot-flasher,
not for U-Boot's command-line. The removed escaping was to work around
the lack of quoting/escaping when passing the user's command to U-Boot's
setenv command during flashing.)

This change will interact badly with the user wanting to use single
quotes in environment variable values, but hopefully that's less likely,
and can be fixed later if needed.

Signed-off-by: Stephen Warren <swarren@nvidia.com>