Use non-zero address for CRC calculations
Currently, U-Boot's itest command cannot read from address 0. On Tegra20,
soc['ram-base'] can be 0, and is the address passed to itest. Modify the
code to use an offset from this address, so as not to use address 0.

Note that in Tegra30 and later, the RAM base is non-zero, so this issue
does not occur. This issue also only affects the "flash" command, and not
the "exec" command.

A patch is also in progress to fix the itest command in U-Boot. However,
fixing tegra-uboot-flasher not to use address 0 seems reasonable too, and
will immediately solve the issue irrespective of which U-Boot code-base is
in use.

Signed-off-by: Stephen Warren <swarren@nvidia.com>